Pass variableValues into the query handler (rather than root args)
Previously the `vars` that were being passed down from the query mock
were in fact the args for the root query selection and therefore already
massaged by the GraphQL executor to have the correct names. This means
that variables to be used further down the graph could either be missing
or have the wrong names.

This is tricky to explain succintly, please see the test case for a
concrete example. In the test case, we are dealing with a scenario where
the variable is used in both the root of the query *and* further down
with a different name. If it were used further down and *not* in the
root query then the variable would simply be absent prior to this patch.

Co-authored-by: Chad Carbert <chadcarbert@me.com>