Drop support for Ember < 3.12
This commit is, I’ll confess, controversial. However, ember-apollo-client
only supports Ember >= 3.4 and it makes our lives *much* easier to follow
suit, and to take that even further and follow ember-cli’s recommendation
to only test the last two LTS releases.

The real source of the 3.4 minimum is ember-native-class-polyfill which
ember-apollo-client now requires Ember < 3.6.

As per [the ember-apollo-client 2.0.0 release notes][1]:

> If you are using Ember versions 3.4 or 3.5 you must add
> ember-native-class-polyfill to your application.

[1]: https://github.com/ember-graphql/ember-apollo-client/releases/tag/v2.0.0